% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_all_missing_by_group.R
\name{count_all_missing_by_group}
\alias{count_all_missing_by_group}
\title{Count missing values for all columns by group}
\usage{
count_all_missing_by_group(data, group_col, .groups = "drop")
}
\arguments{
\item{data}{A data frame or tibble.}

\item{group_col}{Column in the data to group by.}

\item{.groups}{(OPTIONAL) Controls the output data frame's grouping behavior. By default, it drops all levels of grouping ("drop"). Use "keep" to maintain grouping by the specified column. See the \code{summarize()} documentation in \code{dplyr} for additional options.}
}
\value{
A data frame or tibble with the levels of the grouping column and the count of NAs within each level for all columns in the input data except the grouping column.
}
\description{
Given a data frame and a column to group by, this function creates a new data frame with one row per level of the grouping column. The output contains the levels of the grouping column, and the count of missing values for all columns in the input data except the grouping column.
}
\details{
This function uses \code{group_by()} and \code{summarize()} from the \code{dplyr} package to calculate the count of missing values for each group.
}
